[issue-291] extract License and its subclasses to a separate file
