+++ spdx/snippet.py
@@ -10,7 +10,7 @@
 # limitations under the License.
 from typing import Tuple, Optional
 
-from spdx import document
+from spdx import license
 from spdx import utils
 
 
@@ -30,10 +30,10 @@ class Snippet(object):
      which this snippet is associated with. Mandatory, one. Type: str.
      - conc_lics: Contains the license the SPDX file creator has concluded as
      governing the snippet or alternative values if the governing license
-     cannot be determined. Mandatory one. Type: document.License or
+     cannot be determined. Mandatory one. Type: license.License or
      utils.NoAssert or utils.SPDXNone.
      - licenses_in_snippet: The list of licenses found in the snippet.
-     Mandatory, one or more. Type: document.License or utils.SPDXNone or
+     Mandatory, one or more. Type: license.License or utils.SPDXNone or
      utils.NoAssert.
      - attribution_text: optional string.
      - byte_range: Defines the byte range in the original host file that the
@@ -83,7 +83,7 @@ class Snippet(object):
             (str, utils.NoAssert, utils.SPDXNone),
         ):
             messages.append(
-                "Snippet copyright must be str or unicode or utils.NoAssert or utils.SPDXNone"
+                "Snippet copyright must be str or unicode or spdx.utils.NoAssert or spdx.utils.SPDXNone"
             )
 
     def validate_snip_from_file_spdxid(self, messages):
@@ -92,23 +92,23 @@ class Snippet(object):
 
     def validate_concluded_license(self, messages):
         if self.conc_lics and not isinstance(
-            self.conc_lics, (utils.SPDXNone, utils.NoAssert, document.License)
+            self.conc_lics, (utils.SPDXNone, utils.NoAssert, license.License)
         ):
             messages.append(
                 "Snippet concluded license must be instance of "
                 "spdx.utils.SPDXNone or spdx.utils.NoAssert or "
-                "spdx.document.License"
+                "spdx.license.License"
             )
 
     def validate_licenses_in_snippet(self, messages):
         for lic in self.licenses_in_snippet:
             if not isinstance(
-                lic, (document.License, utils.NoAssert, utils.SPDXNone)
+                lic, (license.License, utils.NoAssert, utils.SPDXNone)
             ):
                 messages.append(
                     "License in snippet must be instance of "
                     "spdx.utils.SPDXNone or spdx.utils.NoAssert or "
-                    "spdx.document.License"
+                    "spdx.license.License"
                 )
 
     def has_optional_field(self, field):
